Set header `Content-Type` before setting response status
If this does not get set before setting the response status the header
will not be able to be set.

Also remove unnecessary call to set response status to 500 once the
status code has already been set; this will never get overwritten and
will only be set to 500 if the initial status code was unable to get
set which would indicate an error with the external http package.